requirements.txt
discord.py>=2.4.0
python-dotenv>=1.0.1
pillow>=10.1.0
pyyaml>=6.0.1
aiohttp>=3.9.0
openai>=1.3.8
httpx>=0.25.0
torch>=2.0.0
opencv-python>=4.7.0
pygifsicle>=1.1.0

